refactor: updating components name for Particles
