--- src/components/menu-board/menu-board.module.scss
@@ -1,6 +1,0 @@
-.menuBoard {
-  display: flex;
-  flex-direction: row-reverse;
-  justify-content: space-around;
-  padding: 2rem;
-}
