+++ src/components/menu-boards/menu-board.module.scss
@@ -1,0 +1,6 @@
+.menuBoard {
+  display: flex;
+  flex-direction: row-reverse;
+  justify-content: space-around;
+  padding: 2rem;
+}
